Created statuspage, a health check for a PostgreSQL cluster served using http.
The StatusPage uses cherrypy to serve some simple http endpoints. It returns HTTP Status Code 200
if the endpoint is what it is supposed to be.
The master endpoint will only return 200 if the underlying PostgreSQL is not in recovery.
The slave endpoint will only return 200 if the underlying PostgreSQL is in recovery.

The StatusPage connects to the cluster and sets a nice application name. This connection will be used for executing the queries.